update 02/07/2023 (warning: current for fan need recallibration)

+++ src/sensor-current1/sensor-current1.cpp
@@ -8,7 +8,7 @@ void setupCurrentSensor1() {
 }
 
 void loopCurrentSensor1() {
-    int Irms = (emon1.calcIrms(1480));
+    float Irms = (emon1.calcIrms(1480));
 
     // Assign to global sensing variables
     senseCurrent1 = String(Irms);
